297545: wine-vanilla-7.1-alt2
- set version for provided libwine-devel
- skip linking wine64 to bindir if it is already exists
- don't pack libwinecrt0.a twice
- fix checking for build64mingw, fix build
- fix build without PE
- pack lib*.a needed for build with wine

+++ w/wine-vanilla/wine-vanilla.spec
@@ -6,7 +6,6 @@
 %global __find_debuginfo_files %nil
 %endif
 
-%def_disable static
 %def_with vanilla
 %define gecko_version 2.47.2
 %define mono_version 7.0.0
@@ -26,6 +25,8 @@
 %if_feature llvm 11.0
 # build real PE libraries (.dll, not .dll.so), via clang
 %def_with mingw
+%else
+%def_without mingw
 %endif
 
 # https://bugs.etersoft.ru/show_bug.cgi?id=15244
@@ -62,7 +63,7 @@
 
 Name: wine-vanilla
 Version: %major
-Release: alt1
+Release: alt2
 Epoch: 1
 
 Summary: Wine - environment for running Windows applications
@@ -114,10 +115,9 @@ ExclusiveArch: %ix86 x86_64 aarch64
 
 # TODO: also check build64
 # workaround for https://bugzilla.altlinux.org/38130
-%if_with mingw
-    %def_with build64mingw
-%else
-    %def_without build64mingw
+# notbuild64mingw = without mingw && without build64
+%if %{expand:%%{?_without_mingw:1}%%{!?_without_mingw:0}} && %{expand:%%{?_without_build64:1}%%{!?_without_build64:0}}
+    %define notbuild64mingw 1
 %endif
 
 %define libdir %_libdir
@@ -127,7 +127,6 @@ ExclusiveArch: %ix86 x86_64 aarch64
     %define wineserver wineserver64
     %define winepreloader wine64-preloader
 %else
-    %def_without build64mingw
     %define wineserver wineserver32
     %define winepreloader wine-preloader
 %endif
@@ -485,7 +484,7 @@ Conflicts: %conflictbase-devel-tools
 Conflicts: lib%conflictbase-devel
 Conflicts: lib%name-devel < %version
 %if_without vanilla
-Provides: libwine-devel
+Provides: libwine-devel = %EVR
 %endif
 # we don't need provide anything
 AutoProv:no
@@ -523,19 +522,6 @@ develop programs using %name.
 %name-devel содержит файлы для разработки программ, использующих Wine:
 заголовочные файлы и утилиты, предназначенные
 для компилирования программ с %name.
-
-
-%package devel-static
-Summary: Static libraries for lib%name
-Group: Development/C
-Requires: %name-devel = %EVR
-Conflicts: lib%conflictbase-devel-static
-# we don't need provide anything
-AutoProv:no
-
-%description devel-static
-%name-devel-static contains the static libraries needed to
-develop programs which make use of Wine.
 
 
 %prep
@@ -606,7 +592,7 @@ find %buildroot%winebindir -mindepth 0 -maxdepth 1 -not -type d | \
 # wine64 and wine64-preloader are already built as wine64*
 mv -v %buildroot%winebindir/wineserver %buildroot%winebindir/%wineserver
 %if_with build64
-ln -sv --relative %buildroot%winebindir/wine64 %buildroot%_bindir/
+[ -s %buildroot%_bindir/wine64 ] || ln -sv --relative %buildroot%winebindir/wine64 %buildroot%_bindir/
 %endif
 
 
@@ -648,11 +634,6 @@ rm -rv %buildroot%_mandir/*.UTF-8
 
 # Do not pack dangerous association to run windows executables
 rm -v %buildroot%_desktopdir/wine.desktop
-
-%if_disabled static
-# keep only libwinecrt0.a
-find %buildroot%libwinedir/%winesodir -name "*.a" -not -name libwinecrt0.a -type f -print -delete
-%endif
 
 %if_without debugpe
 # [aarch64] /usr/bin/strip: /usr/src/tmp/wine-staging-buildroot/usr/lib64/wine/aarch64-windows/xinput1_1.dll: file format not recognized
@@ -731,7 +712,6 @@ fi
 %endif
 %libwinedir/%winesodir/winebus.so
 
-# FIXME: https://bugzilla.altlinux.org/38130
 %if_without mingw
 #if_without vanilla
 #libwinedir/%winesodir/windows.networking.connectivity.so
@@ -742,9 +722,11 @@ fi
 %libwinedir/%winesodir/*.ocx.so
 %libwinedir/%winesodir/*.tlb.so
 %libwinedir/%winesodir/*.ax.so
+%libwinedir/%winesodir/*.exe.so
+%libwinedir/%winesodir/*.acm.so
 %endif
 
-%if_without build64mingw
+%ifdef notbuild64mingw
 %libwinedir/%winesodir/*.dll16.so
 %libwinedir/%winesodir/*.drv16.so
 %libwinedir/%winesodir/*.exe16.so
@@ -795,7 +777,7 @@ fi
 %exclude %libwinedir/%winesodir/wined3d.dll.so
 %endif
 
-%if_without build64mingw
+%ifdef notbuild64mingw
 %exclude %libwinedir/%winesodir/twain.dll16.so
 %endif
 
@@ -892,8 +874,10 @@ fi
 %libwinedir/%winesodir/sane.so
 %if_without mingw
 %libwinedir/%winesodir/twain_32.dll.so
+%libwinedir/%winesodir/gphoto2.ds.so
+%libwinedir/%winesodir/sane.ds.so
 %endif
-%if_without build64mingw
+%ifdef notbuild64mingw
 %libwinedir/%winesodir/twain.dll16.so
 %endif
 
@@ -943,19 +927,20 @@ fi
 
 
 %files devel
-#libwinedir/%winesodir/lib*.def
 %if_with mingw
 %libwinedir/%winepedir/lib*.a
 %endif
-%libwinedir/%winesodir/libwinecrt0.a
-
-%if_enabled static
-%files devel-static
 %libwinedir/%winesodir/lib*.a
-%exclude %libwinedir/%winesodir/libwinecrt0.a
-%endif
 
 %changelog
+* Thu Mar 31 2022 Vitaly Lipatov <lav@altlinux.ru> 1:7.1-alt2
+- set version for provided libwine-devel
+- skip linking wine64 to bindir if it is already exists
+- don't pack libwinecrt0.a twice
+- fix checking for build64mingw, fix build
+- fix build without PE
+- pack lib*.a needed for build with wine
+
 * Wed Feb 09 2022 Vitaly Lipatov <lav@altlinux.ru> 1:7.1-alt1
 - new version 7.1 (with rpmrb script)
 - fix build (add hack with ntdll.so and disable debuginfo subpackages)
